door_placement/requirements.txt
shapely>=2.0.0
Pillow>=10.0.0
